# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TestTask",
        "description": "TestTask OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Get user information with orders and its information",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "your_access_token"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "readOnly": "true",
                                            "example": 86400
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "bearer"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "readOnly": "true",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "description": "User unique email address",
                                                    "type": "string",
                                                    "format": "email",
                                                    "readOnly": "true",
                                                    "example": "user@gmail.com"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "readOnly": "true",
                                                    "example": "User Name"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2022-06-11 21:36:35"
                                                },
                                                "finish_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2022-06-11 21:36:35"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "LogOut from the system",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "your_access_token"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "readOnly": "true",
                                            "example": 86400
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "bearer"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "readOnly": "true",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "description": "User unique email address",
                                                    "type": "string",
                                                    "format": "email",
                                                    "readOnly": "true",
                                                    "example": "user@gmail.com"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "readOnly": "true",
                                                    "example": "User Name"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2022-06-11 21:36:35"
                                                },
                                                "finish_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2022-06-11 21:36:35"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh",
                "description": "Refresh users's access token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "your_access_token"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "readOnly": "true",
                                            "example": 86400
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "readOnly": "true",
                                            "example": "bearer"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "readOnly": "true",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "description": "User unique email address",
                                                    "type": "string",
                                                    "format": "email",
                                                    "readOnly": "true",
                                                    "example": "user@gmail.com"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "readOnly": "true",
                                                    "example": "User Name"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2022-06-11 21:36:35"
                                                },
                                                "finish_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2022-06-11 21:36:35"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}